langserve_backend/pyproject.toml
[tool.poetry]
name = "langserve_backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.3.0"}
pydantic = "2.9"
langchain-openai = "^0.2.2"
python-decouple = "^3.8"
langchain-qdrant = "^0.1.4"
qdrant-client = "^1.12.0"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
pymupdf = "^1.24.11"
langgraph = "^0.2.38"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
